+++ NEWS.md
@@ -1,3 +1,20 @@
+<!----------------------------------------------------------------------------->
+<!----------------------------------------------------------------------------->
+## Version 0.4 -- Release notes
+
+### Data model changes
+
+   - new "HPO_parents" table
+   - new "HPO_descendants" table
+   - remove "parent" field from the "HPO_hp" table
+   - new "level" field in the "HPO_hp" table
+
+### Implementation changes
+
+	- Using the "here" package for locating files
+
+<!----------------------------------------------------------------------------->
+<!----------------------------------------------------------------------------->
 ## Version 0.3 -- Release notes
 
 ### Implementation changes